vc-be/pyproject.toml
[project]
name = "vc-be"
version = "0.1.0"
description = "Dating coach backend — FastAPI + LiveKit agent worker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.6",
    "google-genai>=1.0",
    "pinecone>=5.0",
    "livekit>=1.0",
    "livekit-api>=1.0",
    "livekit-agents[google]~=1.4",
]
